feat: added customer create and find by id

+++ domain/src/main/kotlin/repository/CustomerRepository.kt
@@ -3,13 +3,13 @@ package repository
 import model.Customer
 
 interface CustomerRepository {
-    suspend fun create(customer: Customer): Int
-    suspend fun findById(id: Int): Customer?
+    suspend fun create(customer: Customer): Long
+    suspend fun findById(id: Long): Customer?
     suspend fun findByDocument(document: String): Customer?
     suspend fun findAll(): List<Customer>
     suspend fun findActiveCustomers(): List<Customer>
-    suspend fun update(id: Int, customer: Customer): Boolean
-    suspend fun delete(id: Int): Boolean
-    suspend fun softDelete(id: Int): Boolean
-    suspend fun restore(id: Int): Boolean
+    suspend fun update(id: Long, customer: Customer): Boolean
+    suspend fun delete(id: Long): Boolean
+    suspend fun softDelete(id: Long): Boolean
+    suspend fun restore(id: Long): Boolean
 }
